feat: add cors so embedded dashboards can work cross domain
