+++ redash/handlers/static.py
@@ -4,7 +4,7 @@ from flask_login import login_required
 from redash import settings
 from redash.handlers import routes
 from redash.handlers.authentication import base_href
-from redash.handlers.base import org_scoped_rule
+from redash.handlers.base import org_scoped_rule, add_cors_headers
 from redash.security import csp_allows_embeding
 
 
@@ -14,6 +14,9 @@ def render_index():
     else:
         full_path = safe_join(settings.STATIC_ASSETS_PATH, "index.html")
         response = send_file(full_path, **dict(cache_timeout=0, conditional=True))
+
+    if len(settings.ACCESS_CONTROL_ALLOW_ORIGIN) > 0:
+        add_cors_headers(response.headers)
 
     return response
